_bootstrap/app/api/requirements.txt
fastapi
uvicorn
python-decouple